+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.11.0] - 2023-04-03
 ### Added
 - Add support for selecting hypervisor ([#50](https://github.com/cross-platform-actions/action/issues/50))
 - Add support for NetBSD on macOS runners ([#28](https://github.com/cross-platform-actions/action/issues/28))